safe_math

Trait SafeDiv

pub trait SafeDiv {
    // Required methods
    fn safe_div_or(self, rhs: Self, def: Self) -> Self;
    fn safe_div(self, rhs: Self) -> Self;
}

Safe division trait

Required Methods

fn safe_div_or(self, rhs: Self, def: Self) -> Self

Safe division that returns supplied default value for division by zero

fn safe_div(self, rhs: Self) -> Self

Safe division that returns default value for division by zero

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.